requirements.txt
streamlit==1.31.1
ultralytics==8.1.20
google-generativeai==0.3.2
Pillow==10.2.0
opencv-python==4.9.0.80
numpy==1.26.3
torch==2.2.0
torchvision==0.17.0
python-dotenv==1.0.1
